change free memory type to page instead of Mb

+++ kernel/defs.h
@@ -63,7 +63,7 @@ void            ramdiskrw(struct buf*);
 void*           kalloc(void);
 void            kfree(void *);
 void            kinit(void);
-double          kfreespace(void);
+int             kfreespace(void);
 
 // log.c
 void            initlog(int, struct superblock*);
